// examples/next/src/openapi/openapi-typescript/simple.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
	"/pets": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * List all pets
		 * @description List all pets
		 */
		get: operations["listPets"];
		put?: never;
		/**
		 * Create a pet
		 * @description Create a pet
		 */
		post: operations["createPets"];
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
	"/pets/{petId}": {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		/**
		 * Info for a specific pet
		 * @description Info for a specific pet
		 */
		get: operations["showPetById"];
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
}
export type webhooks = Record<string, never>;
export interface components {
	schemas: {
		Pet: {
			/** Format: int64 */
			id: number;
			name: string;
			tag?: string;
		};
		Pets: components["schemas"]["Pet"][];
		Error: {
			/** Format: int32 */
			code: number;
			message: string;
		};
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
	listPets: {
		parameters: {
			query?: {
				/** @description How many items to return at one time (max 100) */
				limit?: number;
			};
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			/** @description An paged array of pets */
			200: {
				headers: {
					/** @description A link to the next page of responses */
					"x-next"?: string;
					[name: string]: unknown;
				};
				content: {
					"application/json": components["schemas"]["Pets"];
				};
			};
			/** @description unexpected error */
			default: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					"application/json": components["schemas"]["Error"];
				};
			};
		};
	};
	createPets: {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			/** @description Null response */
			201: {
				headers: {
					[name: string]: unknown;
				};
				content?: never;
			};
			/** @description unexpected error */
			default: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					"application/json": components["schemas"]["Error"];
				};
			};
		};
	};
	showPetById: {
		parameters: {
			query?: never;
			header?: never;
			path: {
				/** @description The id of the pet to retrieve */
				petId: string;
			};
			cookie?: never;
		};
		requestBody?: never;
		responses: {
			/** @description Expected response to a valid request */
			200: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					"application/json": components["schemas"]["Pets"];
				};
			};
			/** @description unexpected error */
			default: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					"application/json": components["schemas"]["Error"];
				};
			};
		};
	};
}