// packages/openapi-ts/src/types/schema.ts
export interface Extensable {
	[key: `x-${string}`]: any;
}
// ...
export interface ComponentsObject extends Extensable {
	/** An object to hold reusable Schema Objects.*/
	schemas?: Record<string, SchemaObject>;
	/** An object to hold reusable Response Objects. */
	responses?: Record<string, ResponseObject | ReferenceObject>;
	/** An object to hold reusable Parameter Objects. */
	parameters?: Record<string, ParameterObject | ReferenceObject>;
	/** An object to hold reusable Example Objects. */
	examples?: Record<string, ExampleObject | ReferenceObject>;
	/** An object to hold reusable Request Body Objects. */
	requestBodies?: Record<string, RequestBodyObject | ReferenceObject>;
	/** An object to hold reusable Header Objects. */
	headers?: Record<string, HeaderObject | ReferenceObject>;
	/** An object to hold reusable Security Scheme Objects. */
	securitySchemes?: Record<string, SecuritySchemeObject | ReferenceObject>;
	/** An object to hold reusable Link Objects. */
	links?: Record<string, LinkObject | ReferenceObject>;
	/** An object to hold reusable Callback Objects. */
	callbacks?: Record<string, CallbackObject | ReferenceObject>;
	/** An object to hold reusable Path Item Objects. */
	pathItems?: Record<string, PathItemObject | ReferenceObject>;
}
// ...
export interface ParameterObject extends Extensable {
	/**
	 * REQUIRED. The name of the parameter. Parameter names are case sensitive.
	 *
	 * - If `in` is `"path"`, the `name` field MUST correspond to a template expression occurring within the path field in the Paths Object. See Path Templating for further information.
	 * - If `in` is `"header"` and the `name` field is `"Accept"`, `"Content-Type"` or `"Authorization"`, the parameter definition SHALL be ignored.
	 * - For all other cases, the `name` corresponds to the parameter name used by the `in` property.
	 */
	name: string;
	/** REQUIRED. The location of the parameter. Possible values are "query", "header", "path" or "cookie".*/
	in: "query" | "header" | "path" | "cookie";
	/** A brief description of the parameter. This could contain examples of use. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	/** Determines whether this parameter is mandatory. If the parameter location is "path", this property is REQUIRED and its value MUST be true. Otherwise, the property MAY be included and its default value is false. */
	required?: boolean;
	/** Specifies that a parameter is deprecated and SHOULD be transitioned out of usage. Default value is false. */
	deprecated?: boolean;
	/** Sets the ability to pass empty-valued parameters. This is valid only for query parameters and allows sending a parameter with an empty value. Default value is false. If style is used, and if behavior is n/a (cannot be serialized), the value of allowEmptyValue SHALL be ignored. Use of this property is NOT RECOMMENDED, as it is likely to be removed in a later revision. */
	allowEmptyValue?: boolean;
	/** Describes how the parameter value will be serialized depending on the type of the parameter value. Default values (based on value of in): for query - form; for path - simple; for header - simple; for cookie - form. */
	style?: string;
	/** When this is true, parameter values of type `array` or `object` generate separate parameters for each value of the array or key-value pair of the map. For other types of parameters this property has no effect. When `style` is `form`, the default value is `true`. For all other styles, the default value is `false`. */
	explode?: boolean;
	/** Determines whether the parameter value SHOULD allow reserved characters, as defined by [RFC3986] `:/?#[]@!$&'()*+,;=` to be included without percent-encoding. This property only applies to parameters with an `in` value of `query`. The default value is `false`. */
	allowReserved?: boolean;
	/** The schema defining the type used for the parameter. */
	schema?: SchemaObject;
	/** Example of the parameter’s potential value. */
	example?: any;
	/** Examples of the parameter’s potential value. */
	examples?: { [name: string]: ExampleObject | ReferenceObject };
	/** A map containing the representations for the parameter. */
	content?: { [contentType: string]: MediaTypeObject | ReferenceObject };
}

/**
 * [4.8.13] Request Body Object
 * Describes a single request body.
 */
export interface RequestBodyObject extends Extensable {
	/** A brief description of the request body. This could contain examples of use. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	/** REQUIRED. The content of the request body. The key is a media type or media type range and the value describes it. For requests that match multiple keys, only the most specific key is applicable. e.g. text/plain overrides text */
	content: { [contentType: string]: MediaTypeObject | ReferenceObject };
	/** Determines if the request body is required in the request. Defaults to false. */
	required?: boolean;
}
// ...
export interface ResponseObject extends Extensable {
	/** REQUIRED. A description of the response. CommonMark syntax MAY be used for rich text representation. */
	description: string;
	/** Maps a header name to its definition. [RFC7230] states header names are case insensitive. If a response header is defined with the name "Content-Type", it SHALL be ignored. */
	headers?: { [name: string]: HeaderObject | ReferenceObject };
	/** A map containing descriptions of potential response payloads. The key is a media type or media type range and the value describes it. For responses that match multiple keys, only the most specific key is applicable. e.g. text/plain overrides text */
	content?: { [contentType: string]: MediaTypeObject };
	/** A map of operations links that can be followed from the response. The key of the map is a short name for the link, following the naming constraints of the names for Component Objects. */
	links?: { [name: string]: LinkObject | ReferenceObject };
}
// ...
export type CallbackObject = Record<string, PathItemObject>;

/**
 * [4.8.19[ Example Object
 */
export interface ExampleObject extends Extensable {
	/** Short description for the example. */
	summary?: string;
	/** Long description for the example. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	/** Embedded literal example. The value field and externalValue field are mutually exclusive. To represent examples of media types that cannot naturally represented in JSON or YAML, use a string value to contain the example, escaping where necessary. */
	value?: any;
	/** A URI that points to the literal example. This provides the capability to reference examples that cannot easily be included in JSON or YAML documents. The value field and externalValue field are mutually exclusive. See the rules for resolving Relative References. */
	externalValue?: string;
}

/**
 * [4.8.20] Link Object
 * The Link object represents a possible design-time link for a response. The presence of a link does not guarantee the caller’s ability to successfully invoke it, rather it provides a known relationship and traversal mechanism between responses and other operations.
 */
export interface LinkObject extends Extensable {
	/** A relative or absolute URI reference to an OAS operation. This field is mutually exclusive of the operationId field, and MUST point to an Operation Object. Relative operationRef values MAY be used to locate an existing Operation Object in the OpenAPI definition. See the rules for resolving Relative References. */
	operationRef?: string;
	/** The name of an existing, resolvable OAS operation, as defined with a unique operationId. This field is mutually exclusive of the operationRef field. */
	operationId?: string;
	/** A map representing parameters to pass to an operation as specified with operationId or identified via operationRef. The key is the parameter name to be used, whereas the value can be a constant or an expression to be evaluated and passed to the linked operation. The parameter name can be qualified using the parameter location [{in}.]{name} for operations that use the same parameter name in different locations (e.g. path.id). */
	parameters?: { [name: string]: `$${string}` };
	/** A literal value or {expression} to use as a request body when calling the target operation. */
	requestBody?: `$${string}`;
	/** A description of the link. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	/** A server object to be used by the target operation. */
	server?: ServerObject;
}
// ...
export type HeaderObject = Omit<ParameterObject, "name" | "in">;
// ...
export interface ReferenceObject extends Extensable {
	/** REQUIRED. The reference identifier. This MUST be in the form of a URI. */
	$ref: string;
	/** A short summary which by default SHOULD override that of the referenced component. If the referenced object-type does not allow a summary field, then this field has no effect. */
	summary?: string;
	/** A description which by default SHOULD override that of the referenced component. CommonMark syntax MAY be used for rich text representation. If the referenced object-type does not allow a description field, then this field has no effect. */
	description?: string;
}

/**
 * [4.8.24] Schema Object
 * The Schema Object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. This object is a superset of the JSON Schema Specification Draft 2020-12.
 */
export type SchemaObject = {
	/** The Schema Object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. This object is a superset of the JSON Schema Specification Draft 2020-12. */
	discriminator?: DiscriminatorObject;
	/** MAY be used only on properties schemas. It has no effect on root schemas. Adds additional metadata to describe the XML representation of this property. */
	xml?: XMLObject;
	/** Additional external documentation for this schema. */
	externalDocs?: ExternalDocumentationObject;
	/** @deprecated */
	example?: any;
	title?: string;
	description?: string;
	$comment?: string;
	deprecated?: boolean;
	readOnly?: boolean;
	writeOnly?: boolean;
	enum?: unknown[];
	/** Use of this keyword is functionally equivalent to an "enum" (Section 6.1.2) with a single value. */
	const?: unknown;
	default?: unknown;
	format?: string;
	/** @deprecated in 3.1 (still valid for 3.0) */
	nullable?: boolean;
	oneOf?: (SchemaObject | ReferenceObject)[];
	allOf?: (SchemaObject | ReferenceObject)[];
	anyOf?: (SchemaObject | ReferenceObject)[];
	required?: string[];
	[key: `x-${string}`]: any;
} & (
	| StringSubtype
	| NumberSubtype
	| IntegerSubtype
	| ArraySubtype
	| BooleanSubtype
	| NullSubtype
	| ObjectSubtype
	| {
			type: (
				| "string"
				| "number"
				| "integer"
				| "array"
				| "boolean"
				| "null"
				| "object"
			)[];
	  }
	// eslint-disable-next-line @typescript-eslint/ban-types
	| {}
);

export interface TransformObject {
	schema: ts.TypeNode;
	questionToken: boolean;
}

export interface StringSubtype {
	type: "string" | ["string", "null"];
	enum?: (string | ReferenceObject)[];
}
// ...
export type SecuritySchemeObject = {
	/** A description for security scheme. CommonMark syntax MAY be used for rich text representation. */
	description?: string;
	[key: `x-${string}`]: any;
} & (
	| {
			/** REQUIRED. The type of the security scheme. */
			type: "apiKey";
			/** REQUIRED. The name of the header, query or cookie parameter to be used. */
			name: string;
			/** REQUIRED. The location of the API key. */
			in: "query" | "header" | "cookie";
	  }
	| {
			/** REQUIRED. The type of the security scheme. */
			type: "http";
			/** REQUIRED. The name of the HTTP Authorization scheme to be used in the Authorization header as defined in [RFC7235]. The values used SHOULD be registered in the IANA Authentication Scheme registry. */
			scheme: string;
			/** A hint to the client to identify how the bearer token is formatted. Bearer tokens are usually generated by an authorization server, so this information is primarily for documentation purposes. */
			bearer?: string;
	  }
	| {
			/** REQUIRED. The type of the security scheme. */
			type: "mutualTLS";
	  }
	| {
			/** REQUIRED. Tye type of the security scheme. */
			type: "oauth2";
			/** REQUIRED. An object containing configuration information for the flow types supported. */
			flows: OAuthFlowsObject;
	  }
	| {
			/** REQUIRED. Tye type of the security scheme. */
			type: "openIdConnect";
			/** REQUIRED. OpenId Connect URL to discover OAuth2 configuration values. This MUST be in the form of a URL. The OpenID Connect standard requires the use of TLS. */
			openIdConnectUrl: string;
	  }
);